DjangoServer/requirements.txt
django-cors-headers
aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
folium==0.12.1
googlemaps==4.4.5
html5lib
icecream==2.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JPype1
jsonify==0.5
jupyter==1.0.0
konlpy==0.5.2
lxml
MarkupSafe==1.1.1
mccabe
mysql-connector
nltk
numpy~=1.21.5
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5.0.0
python-dateutil==2.8.1
pytz==2020.1
requests~=2.25.1
seaborn
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.4.45
tensorflow~=2.7.0
tqdm~=4.64.1
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
torch~=1.7.1+cu110
django~=4.1
pathlib~=1.0.1
djangorestframework~=3.14.0
matplotlib~=3.5.3
opencv-python~=4.6.0.66
tensorflow-gpu~=2.7.0
scikit-learn~=1.1.3
torchvision~=0.8.2+cu110
dlib~=19.24.0
openface~=0.2.1
ipython~=7.31.1
dataclasses~=0.8
pillow~=9.2.0
tensorflow-datasets~=4.8.0